Extract "disabled message" logic to a hook
