+++ packages/js/src/settings/hooks/index.js
@@ -1,3 +1,4 @@
+export { default as useDisabledMessage } from "./use-disabled-message";
 export { default as useDispatchSettings } from "./use-dispatch-settings";
 export { default as useFormikError } from "./use-formik-error";
 export { default as useRouterScrollRestore } from "./use-router-scroll-restore";
